move save_config
