+++ hydragnn/utils/__init__.py
@@ -18,13 +18,13 @@ from .model import (
     load_existing_model_config,
     loss_function_selection,
     tensor_divide,
-    save_config,
 )
 from .time_utils import Timer, print_timers
 from .config_utils import (
     update_config,
     update_config_minmax,
     get_log_name_config,
+    save_config,
 )
 
 from .optimizer import select_optimizer
